requirements.txt
streamlit
flask
requests
pytesseract
Pillow
